refactor: move twap/containers/TwapConfirmDetails to trade/containers/TradeBasicConfirmDetails
