+++ src/modules/trade/containers/TradeBasicConfirmDetails/styled.ts
